Add variations to the first DNN model for CIFAR-10:
Increase the no. of units of the 2 dense layers in var1.py and var2.py
The result is that accuracy increases, but doubling units in var2 only
improves accuracy by a percent. Also I have not quantified impact on
test data so don't know yet if I am overfitting.

Also that there are other params that I can easily (keep) modify(ing)
and check the impact (in fact an AI can do this itself programatically
as well, i.e. create new static models instead of dynamically adapting,
in the sense of "adding" new units to a layer or adding new layers etc.