Add additional varations to my first CIFAR-10 (non CNN) model
var3, 4 and 5 add more layers while mimicking the number of units
(intuitively by my generational brain model).

var5 proves (from var4) that increasing the number of layers to 5
while maintaining higher unit counts in previous layers does NOT
improve the accuracy to more than 53%

in fact this is lower than var2 which only increased
units in the 2 layers, At the time of writing var5 I had no real
knowledge of measuring the impact to test data to tell if we
were overfitting.