Complete the 4.cnn network and run it.
Results are measly 40 - 43% accuracy. Haven't applied batch normalization
or dropout, and 2 stacked layers instead of 4. The 2 run logs show the results,
and increasing epochs from 10 to 20 doesn't help much.

In the next iteration will play with a couple of optimizers (no LeakyReLU was added
as well to connect the layers).