Go to 75% accuracy : add 4 conv layers, 2 dense layers, activations and batch-normalization, dropout
Model 5.cnn.py on the lines of the ipynb. It takes a bit of experimentation with the middle layers
and with applying batch normalization, dropouts (to prevent overfitting) and activation functions.

While the number of parameters is less than 3.py, accuracy goes up from 49 to 75 percent. This is entirely
due to applying convolutional layers to exploit spacial patterns in the input data, applying variouus activation
functions so the nw learns, applying batch normalizations (to avoid nans, explosive gradients), and dropouts to avoid
overfitting. Model has 4 conv layers, 1 flattened, 2 dense as the middle layers. Layers to be effective must be able
to plugin their inputs and output sizes of the flowing tensors.