Variation: Add 2 hidden convnet layers to CIFAR-10 model
Result is roughly the same accuracy although prediction on test data
improves slightly (mostly in the 70% range). Can see that initial
jumps in accuracy improvements are higher in the earliest epochs and
then the jumps start diminishing. Should the learning rate be increasing
with every epoch if we see diminishing returns in accuracy improvements
per epoch, or are we reaching a true maxima (vs a local one)